Update rule110.py
The rule_110 function has been debugged adding a second list. Doing so, the function does not mutate the same list upon which it evaluates the life-or-death condition of the cell instead, it evaluates an element on the first list and, eventually, changes it in the second list.
Additionally, the rule_110 function in now applied on every following generation through the generator function.